chore(dnsdist): move XskProcessQuery to dnsdist-xsk.cc
